Renames serial_util to swri_serial_util. Refs #231.
